requirements.txt
matplotlib==3.5.1
numpy==1.21.5
pandas==1.3.5
seaborn==0.11.2
tensorboardX==2.6
torchsummary==1.5.1
